sawp: fix dependencies for suricata
Fix dependencies so they are in sync with suricata. Otherwise we get the
following error message.

```
error: failed to select a version for `nom`.
    ... required by package `sawp v0.1.1-rc.2`
    ... which is depended on by `suricata v7.0.0-dev (//suricata/rust)`
versions that meet the requirements `= 5.1.1` are: 5.1.1

all possible versions conflict with previously selected packages.

  previously selected package `nom v5.1.2`
    ... which is depended on by `suricata v7.0.0-dev (//suricata/rust)`

failed to select a version for `nom` which could resolve this conflict
```

Our long term goal is to hide bitflag and nom types from our public API
so there are no version conflicts.