sawp-ike: Add QUICK mode
- IKEv1 may indicate its payloads are encrypted. In this case there is no
available vector of payloads to parse, rather it is an encrypted buffer.
This adds a new field to the IkeMessage which contains the raw encrypted
buffer. This field is only populated when IKEv1 is used and the ENCRYPTED
flag is provided. The vector of payloads is empty in this case. All other
cases (including all IKEv2) will leave encrypted data as None.
- Exposed IkeFlags in API and moved into header.rs.
- Remove raw fields and change FFI to expose them as enums. Removes
dupliation of data while still allowing FFI usage through the enums
rather than raw data.