sawp-ike: Restrict lengths for attribute parsing
When parsing Transforms inside a V1SecurityAssociation
the length of the transform is now used to slice the
input. This allowed for the attribute parser to walk
past the end of the Transform and into the next one,
if there was one.

This often would result in the correct behaviour
accidentally because the final transform header begins
with 0x00 which fails the attribute parser and stops at
the correct spot.

The case it fails is when at least 3 Transforms exist and
the parser walks off the end of the first one, the 0x03
"next payload" field does parse as an attribute and
consumes buffer it should not have had access to.